# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Town {
  name: string;
  /** Day of week the driver picks up here (e.g. "Monday") */
  pickupDay: string;
  dropoffDay: string;
  /**
     * The next scheduled pickup date for this town (YYYY-MM-DD), respecting the midnight-before cutoff
     * @nullable
     */
  nextPickupDate?: string | null;
}

export interface Order {
  id: number;
  orderNumber: string;
  phoneNumber: string;
  name: string;
  town: string;
  colony: string;
  /** @nullable */
  colonyAddress?: string | null;
  unitNumber: string;
  /** @nullable */
  gateAccess?: string | null;
  /** @nullable */
  items?: string | null;
  /** @nullable */
  notes?: string | null;
  /** @nullable */
  cleanerTickets?: string | null;
  /** @nullable */
  pickupDate?: string | null;
  status: string;
  paid: boolean;
  /** @nullable */
  pricedAt?: string | null;
  /** @nullable */
  feeCentsSnapshot?: number | null;
  /** @nullable */
  totalOverrideCents?: number | null;
  totalWasOverridden: boolean;
  /** @nullable */
  receiptSentAt?: string | null;
  /** @nullable */
  paidAt?: string | null;
  /** @nullable */
  paidMethod?: string | null;
  /** @nullable */
  paidConfirmationSentAt?: string | null;
  createdAt: string;
}

export interface CreateOrderInput {
  name: string;
  phoneNumber: string;
  town: string;
  colony: string;
  /** @nullable */
  colonyAddress?: string | null;
  unitNumber: string;
  /** @nullable */
  gateAccess?: string | null;
  /** @nullable */
  items?: string | null;
  /** @nullable */
  notes?: string | null;
  /** @nullable */
  pickupDate?: string | null;
}

/**
 * All fields optional; only provided fields are updated.
 */
export interface UpdateOrderInput {
  name?: string;
  phoneNumber?: string;
  town?: string;
  colony?: string;
  /** @nullable */
  colonyAddress?: string | null;
  unitNumber?: string;
  /** @nullable */
  gateAccess?: string | null;
  /** @nullable */
  items?: string | null;
  /** @nullable */
  notes?: string | null;
  /** @nullable */
  pickupDate?: string | null;
}

export interface OrderStatusUpdate {
  status: string;
}

/**
 * @nullable
 */
export type OrderPaidUpdatePaidMethod = typeof OrderPaidUpdatePaidMethod[keyof typeof OrderPaidUpdatePaidMethod] | null;


export const OrderPaidUpdatePaidMethod = {
  zelle: 'zelle',
  cash: 'cash',
} as const;

export interface OrderPaidUpdate {
  paid: boolean;
  /** @nullable */
  paidMethod?: OrderPaidUpdatePaidMethod;
}

export interface PriceListItem {
  id: number;
  name: string;
  priceCents: number;
  active: boolean;
  sortOrder: number;
  createdAt: string;
  updatedAt: string;
}

export interface PriceListInput {
  name?: string;
  priceCents?: number;
  active?: boolean;
  sortOrder?: number;
}

export interface Settings {
  feeCents: number;
  orderMinimumCents: number;
  wholesalePercent: number;
}

export interface SettingsInput {
  feeCents?: number;
  orderMinimumCents?: number;
  wholesalePercent?: number;
}

export interface OrderLineItem {
  id: number;
  orderId: number;
  /** @nullable */
  priceListId?: number | null;
  itemName: string;
  quantity: number;
  unitPriceCents: number;
  isOverride: boolean;
  sortOrder: number;
  createdAt: string;
}

export interface OrderTotals {
  itemsSubtotalCents: number;
  feeCents: number;
  grandTotalCents: number;
  isOverridden: boolean;
}

export interface OrderPricing {
  lines: OrderLineItem[];
  totals: OrderTotals;
  isPriced: boolean;
}

export interface LineItemInput {
  /** @nullable */
  priceListId?: number | null;
  itemName: string;
  quantity: number;
  unitPriceCents: number;
  isOverride?: boolean;
  sortOrder?: number;
}

export interface ReplaceLineItemsInput {
  lines: LineItemInput[];
  /** @nullable */
  totalOverrideCents?: number | null;
  sendReceipt?: boolean;
}

export interface ReplaceLineItemsResult {
  order: Order;
  lines: OrderLineItem[];
  totals: OrderTotals;
  receiptSent: boolean;
  receiptSkippedReason?: string;
}

export interface EarningsByDay {
  date: string;
  count: number;
  revenueCents: number;
}

export type EarningsReportByMethod = {
  zelle: number;
  cash: number;
  unknown: number;
};

export interface EarningsReport {
  period: string;
  orderCount: number;
  grossRevenueCents: number;
  feesCollectedCents: number;
  itemsRevenueCents: number;
  paidCents: number;
  outstandingCents: number;
  profitEstimateCents: number;
  wholesalePercent: number;
  byMethod: EarningsReportByMethod;
  byRouteDay: EarningsByDay[];
}

export interface BulkMarkReadyResult {
  /** Count of orders moved from at_cleaners to ready */
  updated: number;
  /** The orders that were updated */
  orders: Order[];
}

export type SendOrderReceipt200 = {
  ok: boolean;
};

export type GetEarningsParams = {
period?: GetEarningsPeriod;
};

export type GetEarningsPeriod = typeof GetEarningsPeriod[keyof typeof GetEarningsPeriod];


export const GetEarningsPeriod = {
  today: 'today',
  week: 'week',
  month: 'month',
  all: 'all',
} as const;