# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /orders:
    get:
      operationId: listOrders
      tags: [orders]
      summary: List all dry cleaning orders
      responses:
        "200":
          description: List of orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Order"
    post:
      operationId: createOrder
      tags: [orders]
      summary: Create a new dry cleaning order (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOrderInput"
      responses:
        "201":
          description: Created order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "400":
          description: Validation error
  /orders/{id}:
    patch:
      operationId: updateOrder
      tags: [orders]
      summary: Edit any subset of an order's editable fields
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateOrderInput"
      responses:
        "200":
          description: Updated order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "400":
          description: Validation error
        "404":
          description: Order not found
    delete:
      operationId: deleteOrder
      tags: [orders]
      summary: Permanently delete an order (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Order deleted
        "404":
          description: Order not found
  /orders/{id}/status:
    patch:
      operationId: updateOrderStatus
      tags: [orders]
      summary: Update order status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderStatusUpdate"
      responses:
        "200":
          description: Updated order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "404":
          description: Order not found
...
  /orders/{id}/paid:
    patch:
      operationId: updateOrderPaid
      tags: [orders]
      summary: Update order paid flag and optional payment method
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderPaidUpdate"
      responses:
        "200":
          description: Updated order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "404":
          description: Order not found
  /orders/{id}/line-items:
    get:
      operationId: getOrderLineItems
      tags: [orders]
      summary: Get an order's line items + computed totals
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Order pricing snapshot
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderPricing"
        "404":
          description: Order not found
    put:
      operationId: replaceOrderLineItems
      tags: [orders]
      summary: Replace an order's line items (auto-sends a fresh receipt by default)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReplaceLineItemsInput"
      responses:
        "200":
          description: Updated order + lines + totals
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReplaceLineItemsResult"
        "404":
          description: Order not found
  /orders/{id}/send-receipt:
    post:
      operationId: sendOrderReceipt
      tags: [orders]
      summary: Manually re-send the outstanding receipt SMS
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Sent
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                required: [ok]
        "409":
          description: Cannot send receipt (not priced, or send failed)
...
components:
  schemas:
...
    Order:
      type: object
      properties:
        id:
          type: integer
        orderNumber:
          type: string
        phoneNumber:
          type: string
        name:
          type: string
        town:
          type: string
        colony:
          type: string
        colonyAddress:
          type: ["string", "null"]
        unitNumber:
          type: string
        gateAccess:
          type: ["string", "null"]
        items:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        cleanerTickets:
          type: ["string", "null"]
        pickupDate:
          type: ["string", "null"]
          format: date
        status:
          type: string
        paid:
          type: boolean
        pricedAt:
          type: ["string", "null"]
          format: date-time
        feeCentsSnapshot:
          type: ["integer", "null"]
        totalOverrideCents:
          type: ["integer", "null"]
        totalWasOverridden:
          type: boolean
        receiptSentAt:
          type: ["string", "null"]
          format: date-time
        paidAt:
          type: ["string", "null"]
          format: date-time
        paidMethod:
          type: ["string", "null"]
        paidConfirmationSentAt:
          type: ["string", "null"]
          format: date-time
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - orderNumber
        - phoneNumber
        - name
        - town
        - colony
        - unitNumber
        - status
        - paid
        - totalWasOverridden
        - createdAt
    CreateOrderInput:
      type: object
      properties:
        name:
          type: string
        phoneNumber:
          type: string
        town:
          type: string
        colony:
          type: string
        colonyAddress:
          type: ["string", "null"]
        unitNumber:
          type: string
        gateAccess:
          type: ["string", "null"]
        items:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        pickupDate:
          type: ["string", "null"]
          format: date
      required:
        - name
        - phoneNumber
        - town
        - colony
        - unitNumber
    UpdateOrderInput:
      type: object
      description: All fields optional; only provided fields are updated.
      properties:
        name:
          type: string
        phoneNumber:
          type: string
        town:
          type: string
        colony:
          type: string
        colonyAddress:
          type: ["string", "null"]
        unitNumber:
          type: string
        gateAccess:
          type: ["string", "null"]
        items:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        pickupDate:
          type: ["string", "null"]
          format: date
    OrderStatusUpdate:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    OrderPaidUpdate:
      type: object
      properties:
        paid:
          type: boolean
        paidMethod:
          type: ["string", "null"]
          enum: [zelle, cash, null]
      required:
        - paid
...
    OrderLineItem:
      type: object
      properties:
        id:
          type: integer
        orderId:
          type: integer
        priceListId:
          type: ["integer", "null"]
        itemName:
          type: string
        quantity:
          type: integer
        unitPriceCents:
          type: integer
        isOverride:
          type: boolean
        sortOrder:
          type: integer
        createdAt:
          type: string
          format: date-time
      required: [id, orderId, itemName, quantity, unitPriceCents, isOverride, sortOrder, createdAt]
    OrderTotals:
      type: object
      properties:
        itemsSubtotalCents:
          type: integer
        feeCents:
          type: integer
        grandTotalCents:
          type: integer
        isOverridden:
          type: boolean
      required: [itemsSubtotalCents, feeCents, grandTotalCents, isOverridden]
    OrderPricing:
      type: object
      properties:
        lines:
          type: array
          items:
            $ref: "#/components/schemas/OrderLineItem"
        totals:
          $ref: "#/components/schemas/OrderTotals"
        isPriced:
          type: boolean
      required: [lines, totals, isPriced]
    LineItemInput:
      type: object
      properties:
        priceListId:
          type: ["integer", "null"]
        itemName:
          type: string
        quantity:
          type: integer
        unitPriceCents:
          type: integer
        isOverride:
          type: boolean
        sortOrder:
          type: integer
      required: [itemName, quantity, unitPriceCents]
    ReplaceLineItemsInput:
      type: object
      properties:
        lines:
          type: array
          items:
            $ref: "#/components/schemas/LineItemInput"
        totalOverrideCents:
          type: ["integer", "null"]
        sendReceipt:
          type: boolean
      required: [lines]
    ReplaceLineItemsResult:
      type: object
      properties:
        order:
          $ref: "#/components/schemas/Order"
        lines:
          type: array
          items:
            $ref: "#/components/schemas/OrderLineItem"
        totals:
          $ref: "#/components/schemas/OrderTotals"
        receiptSent:
          type: boolean
        receiptSkippedReason:
          type: string
      required: [order, lines, totals, receiptSent]
...