# lib/api-zod/src/generated/types/order.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Order {
  id: number;
  orderNumber: string;
  phoneNumber: string;
  name: string;
  town: string;
  colony: string;
  /** @nullable */
  colonyAddress?: string | null;
  unitNumber: string;
  /** @nullable */
  gateAccess?: string | null;
  /** @nullable */
  items?: string | null;
  /** @nullable */
  notes?: string | null;
  /** @nullable */
  cleanerTickets?: string | null;
  /** @nullable */
  pickupDate?: Date | null;
  status: string;
  paid: boolean;
  /** @nullable */
  pricedAt?: Date | null;
  /** @nullable */
  feeCentsSnapshot?: number | null;
  /** @nullable */
  totalOverrideCents?: number | null;
  totalWasOverridden: boolean;
  /** @nullable */
  receiptSentAt?: Date | null;
  /** @nullable */
  paidAt?: Date | null;
  /** @nullable */
  paidMethod?: string | null;
  /** @nullable */
  paidConfirmationSentAt?: Date | null;
  createdAt: Date;
}