# lib/api-zod/src/generated/types/orderPaidUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { OrderPaidUpdatePaidMethod } from './orderPaidUpdatePaidMethod';

export interface OrderPaidUpdate {
  paid: boolean;
  /** @nullable */
  paidMethod?: OrderPaidUpdatePaidMethod;
}