# lib/api-zod/src/generated/types/replaceLineItemsInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { LineItemInput } from './lineItemInput';

export interface ReplaceLineItemsInput {
  lines: LineItemInput[];
  /** @nullable */
  totalOverrideCents?: number | null;
  sendReceipt?: boolean;
}