# lib/api-zod/src/generated/types/replaceLineItemsResult.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Order } from './order';
import type { OrderLineItem } from './orderLineItem';
import type { OrderTotals } from './orderTotals';

export interface ReplaceLineItemsResult {
  order: Order;
  lines: OrderLineItem[];
  totals: OrderTotals;
  receiptSent: boolean;
  receiptSkippedReason?: string;
}